backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
openai==0.27.0
pdfminer.six==20201018
python-docx==0.8.11
pymongo==4.6.1
gunicorn==20.1.0
python-magic==0.4.24
spacy==3.1.3
prometheus-client==0.14.1
werkzeug==2.0.1
requests==2.26.0 